workspace/backend/requirements.txt
django>=5.2
djangorestframework>=3.16
psycopg2-binary>=2.9
djangorestframework-simplejwt>=5.5
django-cors-headers>=4.9
python-dotenv>=1.0
whitenoise>=6.5
gunicorn>=21.2